+++ CHANGELOG.md
@@ -6,8 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-- Fix chained input amountIn for `swapETH()` and update `sendTokens()` to `sendFrom()` in `HStargate`.
-- Support Stargate on Fantom
+- Fix chained input amountIn for `swapETH()`.
+- Support Stargate on Fantom.
 
 ## [2.4.0] - 2023-03-02